Tests/LibWeb/Text/expected/input-does-not-scroll-vertically-on-insertion.txt

before insertion: 0
after insertion: 0
vertical inline scrolls: true
vertical block stays stable: true
